requirements.txt
langchain 
langchain-groq 
langchain-core 
neo4j 
sentence-transformers 
transformers 
torch 
python-dotenv 
pandas 
duckduckgo-search 
numpy
accelerate